Git commit message buffers now appear in window that was used to launch them. (#50)
If open_in_new_tab is set to false, when making a git commit using the unception_block_while_host_edits flag, the git commit message buffer will show up in the window of the terminal buffer that was used to run the relevant git command. When the git commit message buffer is quit, the terminal buffer window will be restored. This is more intuitive/consistent with existing functionality for loading in other types of buffers.

Additionally, needing to use :bdelete to unload the commit buffer if 'hidden' is set is no longer necessary. :q should now work, since it triggers the QuitPre event.

Updated demo video.